requirement.txt
selenium
webdriver_manager
python-dotenv
yaspin